添加table指令

+++ go/src/state/api_misc.go
@@ -5,7 +5,7 @@ func (self *luaState) Len(index int) {
 	if s, ok := value.(string); ok {
 		self.stack.push(int64(len(s)))
 	} else if t, ok := value.(*luaTable); ok {
-		self.stack.push(t.len())
+		self.stack.push(int64(t.len()))
 	} else {
 		panic("length error")
 	}
